fix(tools): refactor registerTools to resolve long function issue
